Improve_Eng/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytz>=2023.3
